move DiskManagementError to config-reconcilers
